feat : Monster Data
Monster에 각 Monster마다 다른 데이터 넣는거 시도중, Skill 사용 시 NullReferens 버그 확인

+++ Assets/Resources/Data/MonsterData.cs
@@ -18,7 +18,7 @@ public class MonsterData : ScriptableObject
         public string name;
         public string description;
         
-        public float maxHp;
-        public float haveGeo;
+        public int maxHp;
+        public int haveGeo;
     }
 }
